+++ tests/integration/main.rs
@@ -5,6 +5,7 @@ mod operator_test_utils;
 
 mod test_bson;
 mod test_bytes;
+mod test_cached_object_storage;
 mod test_connector_field_defaults;
 mod test_dd_distinct_total;
 mod test_debezium;